IDEMPIERE-6389: Move test done in AmtInWords_ES.main to the unit tests package
remove main method from AmtInWords_ES
move test class to org.idempiere.test.base
add a test without decimal point
